Refactor `bson2` encoding/decoding (#3987)
